fix: memory leak when decoding Lists of Bytes
Messages that include a `repeated` field of type `bytes` are not
properly freed due to a missing condition `deinit_field()` which only
cares for list of other message types or strings, but not bytes.

This commit introduces a new unit test that fails if decoding a message
with a list of bytes leaks and fix, that makes the unit test pass.