refactor: use switch instead of if condition
In `deinit_field()` we previously used an if condition that checked for
different types within a list, but unfortunately this check was not
exhaustive. By replacing this check with an switch statement, we now
make sure that we have to handle all possible cases, otherwise we'll get
a compiler error.