fix: memory leak when decoding Lists of Bytes (#68)
* fix: memory leak when decoding Lists of Bytes

Messages that include a `repeated` field of type `bytes` are not
properly freed due to a missing condition `deinit_field()` which only
cares for list of other message types or strings, but not bytes.

This commit introduces a new unit test that fails if decoding a message
with a list of bytes leaks and fix, that makes the unit test pass.

* refactor: use switch instead of if condition

In `deinit_field()` we previously used an if condition that checked for
different types within a list, but unfortunately this check was not
exhaustive. By replacing this check with an switch statement, we now
make sure that we have to handle all possible cases, otherwise we'll get
a compiler error.